backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
python-whois==0.8.0
dnspython==2.4.2
firebase-admin==6.2.0
aiosmtplib==2.0.2
python-i18n==0.3.9
cryptography==41.0.5
pytest==7.4.3
httpx==0.25.2
pandas==2.1.3
numpy==1.26.2
tld==0.13
aiohttp==3.9.1
validators==0.22.0 